requirements.txt
# Core dependencies
transformers>=4.35.0
torch>=2.0.0
torchvision>=0.15.0
pillow>=10.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0
gradio>=4.0.0
python-dotenv>=1.0.0
accelerate>=0.20.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0

# Development dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0
mypy>=1.5.0

# Documentation
sphinx>=7.0.0
sphinx-rtd-theme>=1.3.0

# Jupyter (optional, for notebooks)
jupyter>=1.0.0
ipykernel>=6.25.0
ipywidgets>=8.1.0